Update patch set 1
Patch Set 1: Code-Review+1

The reason for this is the new vendor-camera.tegra.so has picture-size-values of 

320x240,480x480,640x368,640x480,800x600,960x960,1088x1088,1024x768,1280x720,1456x1088,2592x1456,2048x1152,640x368,1280x752,1280x960,1440x1080,1536x1536,1600x1200,1836x1080,1920x1080,1952x1952,2048x1152,2048x1360,2048x1536,2448x2448,2592x1456,2592x1520,2592x1920,2592x1944,2592x1952,3264x1840,3264x2448

but android reads them in as 

320x240,480x480,640x368,640x480,800x600,960x960,1088x1088,1024x768,1280x720,1456x1088,2592x1456,2048x1152,640x368,1280x752,1280x960,1440x1080,1536x1536,1600x1200,1836x1080,1920x1080,1952x1952,2048x1152,2048x1360,2048x1536,2448x2448,2592x1456,2592x1520,2592x1920,2592x1944,2592x1952,3

so you lose the higher resolutions. This is exactly 256chars so appears to really be a problem in the android framework reading in >256 char parameters but as the endeavoru and older enrc2b version of the file return the string

320x240,480x480,640x368,640x480,800x600,1024x768,1280x720,1456x1088,2592x1456,2048x1152,640x368,1280x752,1280x960,1440x1080,1600x1200,1836x1080,1920x1080,2048x1152,2048x1360,2048x1536,2592x1456,2592x1520,2592x1920,2592x1944,2592x1952,3264x1840,3264x2448

which is <256 and so works, this fix covers us while still allowing us to use the newer library.

Patch-set: 1