Update patch set 1
Patch Set 1:

About the vibrator stuff. Our drivers are be different and some GPIOs are differently assigned. 

The enrc2b vibrator driver has the following code:

	if(vib->pdata->pwm_gpio >= 0) {
		rc = device_create_file(vib->timed_dev.dev, &dev_attr_dutycycle);
		if (rc){
			pr_err("[VIB] fail to add dutycycle attribute.\n");
			goto err_create_file;
		}
	}

pwm_gpio is set in the board-enrc2b.c file to either -1 or TEGRA_GPIO_PH0 depending on a "board sku tag" and an "pcbid". Not sure if you can enable it by setting it to TEGRA_GPIO_PH0.

Patch-set: 1