Update patch set 1
Patch Set 1:

This change shouldn't affect enrc2b at the moment because you don't set needsHTCCamMode to true in the Gallery2 overlay. But it breaks the endeavoru camera at the moment because of that. Now we could just set needsHTCCamMode to false and call it a day but that would be boring :).

cam-mode 0 is anyway required for some stuff.

Patch-set: 1