Update patch set 2
Patch Set 2:

i don't want to merge this until i can get the kernel to compile again, got a couple of issues i'm ironing out right now. And yes i noticed, i've disabled pollyd in the ramdisk, but thinking of just removing it from the enrc2b stuff, as it's un-needed. pointless having a fix that's un-needed in the rom.

Patch-set: 2