Update patch set 2
Patch Set 2:

The default malloc implementation was changed to jemalloc in Lollipop. This reverts it back to dlmalloc. Google has the same flag in the grouper (also Tegra 3) device tree.

I suspect the reason for this is that there are blobs that were not recompiled for Lollipop and they might not like if the system allocates memory differently than they expect. I'm no expert on that matter so I don't know if that makes sense. But I don't think that it will cause any harm.

Patch-set: 2